+++ CHANGELOG.md
@@ -186,6 +186,7 @@ are doing mtv adjustments during precollision.
 
 ### Updates
 
+- Perf improvement to retrieving components with `ex.Entity.get()` which widely improves engine performance
 - Non-breaking parameters that reference `delta` to `elapsedMs` to better communicate intent and units
 - Perf improvements to `ex.ParticleEmitter` 
   * Use the same integrator as the MotionSystem in the tight loop